Several very minor fixes
Plus the controller for `/servers` endpoints! (No actual logic in the adapter(s), yet, but that's coming next...)

Skipping `/keys` endpoints for the moment because GCE doesn't support adding them that way, technically. I mean, they _might_, in the form of adding the keys to the user's Project, which in this case is the equivalent of a provider account, but it's not clear whether these keys are automatically added to new instances or not, so to play it safe, we'll just assume that's not gonna work, for now...