[Azure] Initial Azure Support
We can't seem to get a location list at the moment, and I can't tell, yet, if it's an Azure permissions issue, or what's going on, but it's a bit of a blocker on generating a catalog. Which means this adapter does virtually nothing useful, yet.

I'm going to dig into Azure and its API a lot more deeply to try to work out the issue and attempt to resolve it.