[Azure] Completed!
Now for some stress testing against Odin.

And yes, we have to generate the password ourselves. Libcloud could generate one for us, but it never actually bothers to report it back to us, so instead we get to generate one which will be guaranteed to be the same every time `GET /server/{id}` is called.

Also lots of error handling in this one, because the Azure driver isn't doing it for us. Or at least not well. It also seems to be using an older version of the MS Azure API, which presents limitations of its own. Hopefully a future version of Libcloud will include an updated Azure driver... (We might need to pitch in to make that happen, at some point, so our users can have access to the full range of available server sizes.)

Did have to add an overridable `_destroy_server()` method, because Azure doesn't clean up after itself on its own like GCE and the others do. It wasn't a huge change, though, thanks to the design already being mostly overridable already.

So, on to the fun part!