[Packet] Mostly Complete
Cannot properly get server creation to work, due to a bug in the Packet driver's handling of `NodeSize`s. Same issue exists when trying to retrieve server info, reboot server, or even delete it.

All other operations work properly, as the driver's implementation can be overridden in those cases. But until the overridden version is merged upstream, the Packet adapter is blocked.