[Azure-Classic] Fixed?
It's got security concerns, probably. But it does work. Successful scale, both up and out.

The trick? Report the external IP for both itself and the internal one, so that both Redd and the firewall connect via the _external_ IP, since the internal IPs can't actually reach each other with the Classic design.

There's a reason it's marked Experimental... But ... it does work...