[Scaleway] Mostly complete
Nearly everything is in place. The only missing element is SSH keys support. That's why the endpoints/values for all three approaches (fully managed, insert on creation, install after start) are still in the OpenAPI spec file - I'm not sure which approach I'll be able to use.

Have some workarounds in mind, and still hoping Scaleway will extend their API to properly support keys natively, but will come back to this tomorrow.

Tonight? One of those.

Oh, also added the ability to change the log level for the Flask side, and fixed a typo.